+++ tests/user-configuration.rs
@@ -17,8 +17,6 @@
 // * along with this program.  If not, see <http://www.gnu.org/licenses/>. *
 // *************************************************************************
 
-#![deny(warnings)]
-
 use std::ffi::CString;
 
 use libc::c_char;