+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 - Adjusted program to use Rust Edition 2018
+- Removed `#![deny(warnings)]` attribute and demoted lints prone to
+  future changes from `deny` to `warn`
 
 
 0.1.2